PP-5167: sqs message reciever scheduled threads test
Validate that correct process name `AND` number of threads are setup and
shutdown by sqs message receiver.
