+++ src/main/java/uk/gov/pay/connector/paymentprocessor/service/CardCaptureMessageProcess.java
@@ -1,14 +1,10 @@
 package uk.gov.pay.connector.paymentprocessor.service;
 
-import com.amazonaws.auth.policy.resources.SQSQueueResource;
-import uk.gov.pay.connector.queue.QueueMessage;
-import uk.gov.pay.connector.queue.sqs.SqsQueueService;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import uk.gov.pay.connector.queue.QueueMessage;
 
 import javax.inject.Inject;
-import java.util.List;
 
 // @TODO(sfount) replace `CardCaptureProcess` when feature flag is switched
 public class CardCaptureMessageProcess {
